Finished converting perlpod to perl kwid.
Everything worked out quite well. afaics, it will be possible to roundtrip pod
and kwid.

Kwid seems very beautiful in comparison to Pod. In the name of
interoperability, Kwid ending up taking on a few podisms, but it does so quite
gracefully.

Autrijus, I removed your '=kwid' line since it is not valid Kwid. It is a Perl
marker to indicate a kwid section. Kwid would simply render it as normal text
in this context.