Updates, clarifications, and expansions to cookbook: introduction to strings
I've also changed to a completely pod format here, and given clearer 
explanation around code. This is to improve upon the previous format I was
using where explanations were too concise and in comments mainly. I'm putting
a note onto p6c shortly to check with the community about this.

Also: if we do stay pod, them I'm sure the formatting and basic structure needs
work if we keep going this direction.

The point is not necessarily to rewrite the book (which I don't own btw, and 
I'm not referencing, but to clarify the specific syntax and options of perl6
that are not clear simple through the code sample's association with the PLEAC
project. Any comments are welcome...this is in the spirit of Wiki editing
through svn :)